+++ libs/compat/CHANGELOG.rst
@@ -1,3 +1,7 @@
+**0.2.1 - 05/14/26**
+
+- Update test_removal_deadline to reflect new removal deadline of 2027-07-01.
+
 **0.2.0 - 05/13/26**
 
 - Activate ``layered_config_tree`` -> ``vivarium.config_tree`` redirect.